+++ src/php/Gdbots/Schemas/Ncr/Mixin/IndexNode/IndexNodeV1Mixin.php
@@ -24,7 +24,7 @@ final class IndexNodeV1Mixin extends AbstractMixin
     {
         return [
             Fb::create('node', T\MessageType::create())
-                ->className('Gdbots\Schemas\Ncr\Mixin\Node\Node')
+                ->className('Gdbots\Schemas\Ncr\Mixin\Indexed\Indexed')
                 ->build()
         ];
     }