+++ src/php/Gdbots/Schemas/Ncr/Mixin/Indexable/IndexableV1.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Gdbots\Schemas\Ncr\Mixin\Indexable;
-
-interface IndexableV1 extends Indexable
-{
-}