+++ src/php/Gdbots/Schemas/Ncr/Mixin/Indexed/IndexedV1Mixin.php
@@ -1,20 +1,20 @@
 <?php
 
-namespace Gdbots\Schemas\Ncr\Mixin\Indexable;
+namespace Gdbots\Schemas\Ncr\Mixin\Indexed;
 
 use Gdbots\Pbj\AbstractMixin;
 use Gdbots\Pbj\FieldBuilder as Fb;
 use Gdbots\Pbj\SchemaId;
 use Gdbots\Pbj\Type as T;
 
-final class IndexableV1Mixin extends AbstractMixin
+final class IndexedV1Mixin extends AbstractMixin
 {
     /**
      * {@inheritdoc}
      */
     public function getId()
     {
-        return SchemaId::fromString('pbj:gdbots:ncr:mixin:indexable:1-0-0');
+        return SchemaId::fromString('pbj:gdbots:ncr:mixin:indexed:1-0-0');
     }
 
     /**