+++ src/php/Gdbots/Schemas/Ncr/Mixin/Indexed/IndexedV1Trait.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Gdbots\Schemas\Ncr\Mixin\Indexed;
+
+use Gdbots\Pbj\Schema;
+
+/**
+ * @method static Schema schema
+ * @method mixed get($fieldName, $default = null)
+ */
+trait IndexedV1Trait
+{
+}