+++ src/php/Gdbots/Schemas/Ncr/Mixin/NodePublished/NodePublishedV1Mixin.php
@@ -30,7 +30,6 @@ final class NodePublishedV1Mixin extends AbstractMixin
                 ->format(Format::SLUG())
                 ->build(),
             Fb::create('published_at', T\DateTimeType::create())
-                ->useTypeDefault(false)
                 ->build()
         ];
     }