+++ src/php/Gdbots/Schemas/Ncr/Mixin/NodeScheduled/NodeScheduledV1Mixin.php
@@ -30,7 +30,6 @@ final class NodeScheduledV1Mixin extends AbstractMixin
                 ->format(Format::SLUG())
                 ->build(),
             Fb::create('publish_at', T\DateTimeType::create())
-                ->useTypeDefault(false)
                 ->build()
         ];
     }