+++ src/php/Gdbots/Schemas/Ncr/Mixin/Publishable/PublishableV1Mixin.php
@@ -24,7 +24,6 @@ final class PublishableV1Mixin extends AbstractMixin
     {
         return [
             Fb::create('published_at', T\DateTimeType::create())
-                ->useTypeDefault(false)
                 ->build()
         ];
     }