+++ src/php/Gdbots/Schemas/Pbjx/Mixin/Indexed/Indexed.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Gdbots\Schemas\Pbjx\Mixin\Indexed;
+
+use Gdbots\Schemas\Pbjx\Mixin\Event\EventV1;
+
+interface Indexed extends EventV1
+{
+}