+++ src/php/Gdbots/Schemas/Pbjx/Mixin/Indexed/IndexedV1.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Gdbots\Schemas\Pbjx\Mixin\Indexed;
+
+interface IndexedV1 extends Indexed
+{
+}