+++ src/php/Gdbots/Schemas/Pbjx/Mixin/Indexed/IndexedV1Mixin.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace Gdbots\Schemas\Pbjx\Mixin\Indexed;
+
+use Gdbots\Pbj\AbstractMixin;
+use Gdbots\Pbj\FieldBuilder as Fb;
+use Gdbots\Pbj\SchemaId;
+use Gdbots\Pbj\Type as T;
+
+final class IndexedV1Mixin extends AbstractMixin
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function getId()
+    {
+        return SchemaId::fromString('pbj:gdbots:pbjx:mixin:indexed:1-0-0');
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getFields()
+    {
+        return [];
+    }
+}