+++ src/php/Gdbots/Schemas/Pbjx/Mixin/SearchEventsRequest/SearchEventsRequest.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Gdbots\Schemas\Pbjx\Mixin\SearchEventsRequest;
+
+use Gdbots\Schemas\Pbjx\Mixin\Request\RequestV1;
+
+interface SearchEventsRequest extends RequestV1
+{
+}