+++ src/php/Gdbots/Schemas/Pbjx/Mixin/SearchEventsRequest/SearchEventsRequestV1Trait.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Gdbots\Schemas\Pbjx\Mixin\SearchEventsRequest;
+
+use Gdbots\Pbj\Schema;
+use Gdbots\Schemas\Pbjx\Enum\SearchSort;
+
+/**
+ * @method static Schema schema
+ * @method mixed get($fieldName, $default = null)
+ */
+trait SearchEventsRequestV1Trait
+{
+}